moved canopen::Settings into can namespace
